requirements.txt
aiogram
python-dotenv
peewee